requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
psycopg[binary]==3.2.3
pydantic==2.9.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
